requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
httpx>=0.27.0
